+++ Electricity_LCI_EPA_Emissions-master/.gitignore
@@ -1,0 +1,5 @@
+*.xlsx 
+*.csv
+
+electricitylci/generationtemplate(old).py
+electricitylci/generationtemplate_(old).py
