+++ Electricity_LCI_EPA_Emissions-master/Distribution-Mix-master/.gitattributes
@@ -1,0 +1,2 @@
+# Auto detect text files and perform LF normalization
+* text=auto
