+++ Electricity_LCI_EPA_Emissions-master/Generation/.gitignore
@@ -1,0 +1,2 @@
+
+*.csv
