+++ Electricity_LCI_EPA_Emissions-master/Generation/globals.py
