+++ Electricity_LCI_EPA_Emissions-master/Generation/old files/globals.py
@@ -1,0 +1,8 @@
+import os
+def set_dir(directory):
+    if not os.path.exists(directory): os.makedirs(directory)
+    return directory
+
+
+datadir = 'electricitylci/data/'
+outputdir = set_dir('output/')
