--- electricitylci/REFACTOR/processes/context.json
@@ -1,1 +1,0 @@
-{"@vocab":"http://openlca.org/schema/v1.0/","@base":"http://openlca.org/schema/v1.0/","modelType":{"@type":"@vocab"},"flowPropertyType":{"@type":"@vocab"},"flowType":{"@type":"@vocab"},"distributionType":{"@type":"@vocab"},"parameterScope":{"@type":"@vocab"},"allocationType":{"@type":"@vocab"},"defaultAllocationMethod":{"@type":"@vocab"},"allocationMethod":{"@type":"@vocab"},"processType":{"@type":"@vocab"},"riskLevel":{"@type":"@vocab"}}
