--- electricitylci/REFACTOR/processes/data.json
@@ -1,1 +1,0 @@
-{"allocationFactors": "", "defaultAllocationMethod": "", "exchanges": [{"": "", "internalId": "", "avoidedProduct": "False", "flow": {"": "", "flowType": "PRODUCT_FLOW", "cas": "biomass", "formula": "", "flowProperties": "", "location": "AZNM"}, "flowProperty": "", "input": "True", "quantitativeReference": "True", "baseUncertainty": "", "provider": "", "amount": NaN, "amountFormula": "", "unit": "kg", "pedigreeUncertainty": "", "uncertainty": {"": "", "distributionType": "Logarithmic Normal Distribution", "mean": "", "meanFormula": "", "geomMeanFormula": "", "minimum": 2985608045.0, "minimumFormula": "", "sd": "", "sdFormula": "", "geomSdFormula": "", "mode": "", "modeFormula": "", "maximum": 2985608045.0, "maximumFormula": ""}, "comment": "eGRID 2014"}, {"": "", "internalId": "", "avoidedProduct": "False", "flow": {"": "", "flowType": "PRODUCT_FLOW", "cas": "Carbon dioxide", "formula": "", "flowProperties": "", "location": "AZNM"}, "flowProperty": "", "input": "False", "quantitativeReference": "", "baseUncertainty": "", "provider": "", "amount": 2.5090746860961586, "amountFormula": "", "unit": "kg", "pedigreeUncertainty": "", "uncertainty": {"": "", "distributionType": "Logarithmic Normal Distribution", "mean": "", "meanFormula": "", "geomMeanFormula": "", "minimum": 426809.0, "minimumFormula": "", "sd": "", "sdFormula": "", "geomSdFormula": "", "mode": "", "modeFormula": "", "maximum": 426809.0, "maximumFormula": ""}, "comment": "eGRID 2014"}], "location": "AZNM", "parameters": "", "processDocumentation": {"": "", "timeDescription": "", "validUntil": "12/312014", "validFrom": "1/1/2014", "technologyDescription": "", "dataCollectionDescription": "", "completenessDescription": "", "dataSelectionDescription": "", "reviewDetails": "", "dataTreatmentDescription": "", "inventoryMethodDescription": "", "modelingConstantsDescription": "", "reviewer": "Wes Ingwersen", "samplingDescription": "", "sources": "", "restrictionsDescription": "", "copyright": "", "creationDate": "", "dataDocumentor": "Wes Ingwersen", "dataGenerator": "Tapajyoti Ghosh", "dataSetOwner": "", "intendedApplication": "", "projectDescription": "", "publication": "", "geographyDescription": ""}, "processType": ""}
