--- stewi/data/DMR/2_digit_SIC.txt
@@ -1,84 +1,0 @@
-
-01 - Agricultural Production, Crops
-02 - Agricultural Production, Livestock
-07 - Agricultural Services
-08 - Forestry
-09 - Fishing, Hunting, & Trapping
-10 - Metal Mining
-12 - Coal Mining
-13 - Natural Gas Liquids
-14 - Mining And Quarrying Of Nonmetallic Minerals, Except Fuels
-15 - General Building Contractors
-16 - Heavy Construction, Except Building
-17 - Special Trade Contractors
-20 - Food & Kindred Products
-21 - Tobacco Products
-22 - Textile Mill Products
-23 - Apparel & Other Textile Products
-24 - Lumber & Wood Products
-25 - Furniture & Fixtures
-26 - Paper & Allied Products
-27 - Printing & Publishing
-28 - Chemical & Allied Products
-29 - Petroleum & Coal Products
-30 - Rubber And Miscellaneous Plastics Products
-31 - Leather & Leather Products
-32 - Stone, Clay, & Glass Products
-33 - Primary Metal Industries
-34 - Fabricated Metal Products
-35 - Industrial And Commercial Machinery And Computer Equipment
-36 - Electronic And Other Electrical Equipment And Components, Ex
-37 - Transportation Equipment
-38 - Measuring, Analyzing, and Controlling Instruments
-39 - Misc. Manuf. Industries
-40 - Railroad Transportation
-41 - Local & Interurban Passenger Transitoption
-42 - Trucking & Warehousing
-43 - U.S. Postal Service
-44 - Water Transportation
-45 - Transportation By Air
-46 - Pipelines, Except Natural Gas
-47 - Transportation Services
-48 - Communications
-49 - Electric, Gas, & Sanitary Services
-50 - Wholesale Trade, Durable Goods
-51 - Wholesale Trade, Nondurable Goods
-52 - Building Materials & Gardening Supplies
-53 - General Merchandise Stores
-54 - Food Stores
-55 - Automative Dealers & Service Stations
-56 - Apparel & Accessory Stores
-57 - Furniture & Homefurnishings Stores
-58 - Eating & Drinking Places
-59 - Miscellaneous Retail
-60 - Depository Institutions
-61 - Nondepository Institutions
-62 - Security & Commodity Brokers
-63 - Insurance Carriers
-64 - Insurance Agents, Brokers, & Service
-65 - Real Estate
-67 - Holding & Other Investment Offices
-70 - Hotels & Other Lodging Places
-72 - Personal Services
-73 - Business Services
-75 - Auto Repair, Services, & Parking
-76 - Miscellaneous Repair Services
-78 - Motion Pictures
-79 - Amusement & Recreation Services
-80 - Health Services Other Than Hospitals
-81 - Legal Services
-82 - Educational Services
-83 - Social Services
-84 - Museums, Botanical, Zoological Gardens
-86 - Membership Organizations
-87 - Engineering & Management Services
-88 - Private Households
-89 - Services, Not Elsewhere Classified
-91 - Executive, Legislative, & General
-92 - Justice, Public Order, & Safety
-93 - Finance, Taxation, & Monetary Policy
-95 - Environmental Quality & Housing
-96 - Administration Of Economic Programs
-97 - National Security & International Affairs
-99 - Non Classifiable Establishments
-
