--- stewi/data/DMR_required_fields.txt
@@ -1,13 +1,0 @@
-EstFactor
-ExternalPermitNmbr
-FacilityTypeIndicator
-NaicsCode
-ParameterDesc
-PollutantLoad
-Siccode
-StateCode
-TRIDirectPounds
-TRIFids
-TRIIndirectPounds
-Uin
-
