--- stewi/data/RCRAImportSourceCodes.txt
@@ -1,14 +1,0 @@
-G61
-G63
-G64
-G65
-G66
-G67
-G68
-G69
-G70
-G71
-G72
-G73
-G74
-G75
