--- stewi/data/RCRA_required_fields.txt
@@ -1,19 +1,0 @@
-Handler ID
-State
-Generator ID Included in NBR
-Generator Waste Stream Included in NBR
-Waste Description
-Primary NAICS
-Source Code
-Form Code
-Management Method
-Generation Tons
-Federal Waste Flag
-Waste Code Group
-
-
-
-
-
-
-
