--- stewi/data/TRI_required_fields.txt
@@ -1,35 +1,0 @@
-TRIFID
-FACILITY NAME
-FACILITY STREET
-FACILITY CITY
-FACILITY COUNTY
-FACILITY STATE
-FACILITY ZIP CODE
-PRIMARY NAICS CODE
-LATITUDE
-LONGITUDE
-CAS NUMBER
-CHEMICAL NAME
-UNIT OF MEASURE
-TOTAL FUGITIVE AIR EMISSIONS
-FUGITIVE OR NON-POINT AIR EMISSIONS - BASIS OF ESTIMATE
-TOTAL STACK AIR EMISSIONS
-STACK OR POINT AIR EMISSIONS - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM A
-DISCHARGES TO STREAM A - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM B
-DISCHARGES TO STREAM B - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM C
-DISCHARGES TO STREAM C - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM D
-DISCHARGES TO STREAM D - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM E
-DISCHARGES TO STREAM E - BASIS OF ESTIMATE
-TOTAL DISCHARGES TO STREAM F
-DISCHARGES TO STREAM F - BASIS OF ESTIMATE
-TOTAL LAND TREATMENT
-LAND TRTMT/APPL FARMING - BASIS OF ESTIMATE
-TOTAL OTHER DISPOSAL
-OTHER DISPOSAL -BASIS OF ESTIMATE
-LAND TREATMENT
-OTHER LAND DISPOSAL
