--- stewi/data/egrid_required_fields.txt
@@ -1,22 +1,0 @@
-Plant state abbreviation
-Plant name
-Plant operator name
-DOE/EIA ORIS plant or facility code
-Plant total annual heat input (MMBtu)
-Plant annual net generation (MWh)
-CHP plant useful thermal output (MMBtu)
-Plant annual NOx emissions (tons)
-Plant annual SO2 emissions (tons)
-Plant annual CO2 emissions (tons)
-Plant annual CH4 emissions (lbs)
-Plant annual N2O emissions (lbs)
-Plant primary fuel
-Plant nameplate capacity (MW)
-Plant primary coal/oil/gas/ other fossil fuel category
-Prime Mover
-NERC region acronym
-eGRID subregion acronym
-Balancing Authority Name
-Plant county name
-Plant latitude
-Plant longitude
