--- stewi/data/flowbyfacility_optional_output_fields.json
@@ -1,9 +1,0 @@
-{
-	"TRI": ["Compartment"],
-	"NEI": "NA",
-	"RCRA_Info": "NA",
-    "GHGRP": "NA",
-    "DMR": "NA",
-	"eGRID": "NA",
-	"RCRAInfo": "NA"
-}
