--- stewi/output/RCRAInfo_2013_metadata.json
@@ -1,1 +1,0 @@
-{"SourceType": "Static File", "SourceFileName": "../RCRAInfo/br_reporting_2013.txt", "SourceURL": "ftp://ftp.epa.gov/rcrainfodata/rcra_flatfiles/Baseline/biennial_report.zip", "SourceVersion": "NA", "SourceAquisitionTime": "Sat Apr 28 22:13:01 2018", "StEWI_versions_version": "0.9"}
