--- stewi/output/eGRID_2016_metadata.json
@@ -1,1 +1,0 @@
-{"SourceType": "Static File", "SourceFileName": "../eGRID/egrid2016_Data.xlsx", "SourceURL": "https://www.epa.gov/sites/production/files/2018-02/egrid2016_all_files_since_1996.zip", "SourceVersion": "", "SourceAquisitionTime": "Wed May 10 10:00:01 2018", "StEWI_versions_version": "0.9"}
